Directly use BigUint64 for bitset

+++ native.d.ts
@@ -4,7 +4,7 @@ export declare function parse(code: string, allowReturnOutsideFunction: boolean,
 
 export declare function parseAsync(code: string, allowReturnOutsideFunction: boolean, jsx: boolean, signal?: AbortSignal | undefined | null): Promise<Buffer>
 
-export declare function parseAndWalk(code: string, allowReturnOutsideFunction: boolean, jsx: boolean, walkedNodesBitset: Uint8Array, signal?: AbortSignal | undefined | null): Promise<Buffer>
+export declare function parseAndWalk(code: string, allowReturnOutsideFunction: boolean, jsx: boolean, walkedNodesBitset: BigUint64Array, signal?: AbortSignal | undefined | null): Promise<Buffer>
 
 export declare function xxhashBase16(input: Uint8Array): string
